Code docs and run->run_and_measure
Fixed documentation so example in the qaoa installation and quickstart is
correct.

run_and_measure is now used instead of run which significantly reduces the
amount of time required to sample a circuit.  This method speeds up simulation
by not re-running the circuit for every sample.  It runs the circuit once,
stores the wavefunction, and then does the projection step `samples` number of
times.