Finding active qubits in a list of pauli terms
There was an error in how the map from qubit to measurement result was
constructed.

Old code would have taken a list of Pauli terms consisting of

```
(-0.019844113673352048+0j)*X0*X1*X3*Z4*Z5*X6 + (-0.4047216515044083+0j)*Z2 + (-1.0378580892109397+0j)*Z7
```

and returned the `active_qubit_indices` as [0, 1, 3, 4, 5, 6] because it
was erroneously looking for the maximal weight term instead of the union
of all qubits being examined.