Easy estimation (#151)
* Improved interface for operator averaging

Improved interface for operator averaging with a term grouping such that
the diagonal terms are grouped together.  We can probably generalize
this so that we take a method for grouping and return the estimator.

* Check empty identity term

so this works with all pauli sums

* Method changes and testing updates

Changed method name to something sane and fixed failing test.  Test was
failing because qubit ordering has now changed.  It should be noted that
many other tests are likely to fail now.

* Reverting test so it works with pyquil 1.8.0

The latest release does not contain significant reordering changes that
were introduced into the pyquil pauli operators.  Since semaphore
testing happens against the latest release (instead of master) a
rotation test failed.  I am reverting the ordering of the test so it
passes.

* remove positivity constriant change

Make another PR for this change/open an issue for discussion